==> Part1.dataset_cleaning_and_merging.ipynb <==
### Dataset cleaning and merging

#### 1) In this step the 3 lung cancer datasets are stripped off of both common and non-common elements.
#### 2) This is done in an effort to preserve space and to omit having NaN values due to non-common elements.
#### 3) Subsequently the 3 datasets are concatenated into a single anndata object and the tissue values are replaced (see dictionary below for details).

import scanpy as sc
import anndata as ad
import pandas as pd
import hdf5plugin

### Parameters

# nsclc_tissue_dic = {
#     "tumor": "cancer",
#     "blood": "normal",
#     "adjacent normal": "normal",
# }


# luad_gse131907_tissue_dic = {
#     "tLung": "cancer",
#     "nLung": "cancer",
#     "nLN": "cancer",
#     "mBrain": "cancer",
#     "mLN": "cancer",
#     "PE": "cancer",
#     "tL/B": "cancer",
# }


# luad_gse123902_tissue_dic = {
#     'PRIMARY': "cancer",
#     'NORMAL': "normal",
#     'METASTASIS': "cancer",
# }


luad_gse97168_tissue_dic = {
    'normal': "normal",
    'tumor': "cancer",
}

### Dataset paths

# DATA_PATH = "./data/standalone_h5ads/NSCLC_T_SS2_GSE99254.h5ad"

# DATA_PATH = "./data/standalone_h5ads/LUAD_UNB_10X_GSE131907.h5ad"

# DATA_PATH = "./data/standalone_h5ads/LUAD_UNB_10X_GSE123902.h5ad"

DATA_PATH = "./data/standalone_h5ads/LUAD_MYE_MRS_GSE97168.h5ad"

### Read & Sanitize adata object

adata = sc.read_h5ad(DATA_PATH)
adata

# Step 1: Keep only specific columns in `.obs`
# Replace 'column1', 'column2' with the names of the columns you want to keep
adata.obs = adata.obs[['tissue', 'annotation_immune']]

# Step 2: Remove all columns from `.var`
# This retains the gene names/index but removes all associated annotations
adata.var = pd.DataFrame(index=adata.var.index)

# Step 3: Remove all items from `.obsm`
adata.obsm = {}

adata.obs["tissue"].value_counts()

adata.obs.rename(columns = {'Tissue':'tissue'}, inplace = True)
adata.obs['tissue'] = adata.obs['tissue'].replace(luad_gse97168_tissue_dic)

adata.obs["tissue"].value_counts()

### Write sanitized dataset to disk

adata.write_h5ad(
    "./data/sanitized_h5ads/luad_gse97168.h5ad",
    compression=hdf5plugin.FILTERS["zstd"],
    compression_opts=hdf5plugin.Zstd(clevel=5).filter_options
                )
