==> Part3.scvi_solo_doublet_identification.ipynb <==
import scanpy as sc
import pandas as pd
import matplotlib.pyplot as plt
import scvi
import hdf5plugin
import os

def filter_doublets(directory="./data/3_filtered_h5ads"):
    # List all files in the specified directory
    for file in os.listdir(directory):
        if file.endswith(".h5ad"):
            # Construct the full file path
            file_path = os.path.join(directory, file)
            # Print the file name
            print(f"Processing file: {file_path}")

            adata = sc.read_h5ad(file_path)
            sc.pp.highly_variable_genes(adata, n_top_genes=3000, subset=True, flavor="seurat", span=0.8)
            scvi.model.SCVI.setup_anndata(adata)
            vae = scvi.model.SCVI(adata)
            vae.train()
            solo = scvi.external.SOLO.from_scvi_model(vae)
            solo.train()
            df = solo.predict()
            df["prediction"] = solo.predict(soft=False)

            doublet_dic = dict(zip(df.index, df.prediction))

            def filter_doublet(x):
                try:
                    return doublet_dic[x]
                except:
                    return 'filtered'

            adata = sc.read_h5ad(file_path)
            adata.obs["doublet"] = adata.obs.index.map(filter_doublet)
            adata = adata[adata.obs.doublet == 'singlet']
            adata.write_h5ad(
                        f"./data/4_fwd_h5ads/fwd_{file}",
                        compression=hdf5plugin.FILTERS["zstd"]
                    )

filter_doublets()
