==> Part4.concatenate_data.ipynb <==
import scanpy as sc
import hdf5plugin
import anndata
import matplotlib.pyplot as plt
from matplotlib.pyplot import rc_context
import numpy as np
import pandas as pd
import os

import warnings
warnings.filterwarnings("ignore")

VERSION = "3"

DATA_PATH = "./data/2_sanitized_h5ads"

def get_h5ads(directory):
    h5ad_name_lst = []
    for file in os.listdir(directory):
        if file.endswith(".h5ad"):
            # Construct the full file path
            file_path = os.path.join(directory, file)
            # Print the file name
            h5ad_name_lst.append(file)
    return h5ad_name_lst

adata_name_lst = get_h5ads(DATA_PATH)
adata_name_lst

H5ADS = [f"{DATA_PATH}/{i}" for i in adata_name_lst]

MERGED_NAME = "lung_cancer_concat_data"
H5AD_CONCAT = f"./data/4_concatenated_h5ad/{MERGED_NAME}_v{VERSION}.h5ad"

H5ADS

adatas = [sc.read_h5ad(i) for i in H5ADS]
adatas

adata_merged = anndata.AnnData.concatenate(*adatas, batch_key='batch', join="inner")
adata_merged.obs.rename(columns={'annotation_immune': 'cell_type'}, inplace=True)
adata_merged

adata_merged.write_h5ad(
    H5AD_CONCAT,
    compression=hdf5plugin.FILTERS["zstd"],
    compression_opts=hdf5plugin.Zstd(clevel=5).filter_options
)
