==> Part4.2.adata2dataframe.ipynb <==
import hdf5plugin
import scanpy as sc
import numpy as np
import pandas as pd

# ADATA_PATH = "./data/4_concatenated_h5ad/lung_cancer_concat_data_v2.h5ad"

ADATA_PATH = "./data/4_concatenated_h5ad/lung_cancer_concat_data_v3.h5ad"

merged_adata = sc.read_h5ad(ADATA_PATH)
merged_adata

sc.pp.normalize_total(merged_adata, target_sum=1e4)
sc.pp.log1p(merged_adata)
sc.pp.scale(merged_adata, max_value=10)

df = pd.DataFrame(merged_adata.X)
df

df.columns = merged_adata.var.index
df

df["tissue"] = merged_adata.obs["tissue"].values
df["cell_type"] = merged_adata.obs["cell_type"].values

df["cell_type"].value_counts()

merged_adata.obs["cell_type"].value_counts()

# Reset the index and do not add the old index as a column
df = df.reset_index(drop=True)
df

# df.to_csv("./data/5_csv_dataframes/lc_concatenated_df.csv", index=False)

# To save the DataFrame as a pickle file
df.to_pickle("./data/5_csv_dataframes/lc_concatenated_df_v3.pkl")
