==> Part5.multilabel_classification.ipynb <==
from sklearn.ensemble import RandomForestClassifier
from sklearn.multioutput import MultiOutputClassifier
from sklearn.utils import shuffle
import pandas as pd
import numpy as np

# When loading, you'll get the DataFrame without the original index
# df = pd.read_pickle("./data/5_csv_dataframes/lc_concatenated_df.pkl")

# df = pd.read_pickle("./data/5_csv_dataframes/lc_concatenated_df_v2.pkl")

df = pd.read_pickle("./data/5_csv_dataframes/lc_concatenated_df_v3.pkl")
df

df['cell_type_code'] = df['cell_type'].astype('category').cat.codes


condition_mapping = {'normal': 0, 'cancer': 1}
df['condition_code'] = df['tissue'].map(condition_mapping)

df

X = df.drop(['cell_type', 'tissue', 'cell_type_code', 'condition_code'], axis=1)
y = df[['cell_type_code', 'condition_code']]

X

rfc = RandomForestClassifier(n_estimators=100, random_state=42)

multi_target_rfc = MultiOutputClassifier(rfc, n_jobs=-1)

multi_target_rfc.fit(X, y)

# Feature importance for each label
importances = multi_target_rfc.estimators_[0].feature_importances_
feature_names = X.columns
importances

# Create a DataFrame for feature importances
feature_importances = np.mean([tree.feature_importances_ for tree in multi_target_rfc.estimators_], axis=0)

features_df = pd.DataFrame({
    'feature': X.columns,
    'importance': feature_importances
})

# Sort the DataFrame by importance
features_df_sorted = features_df.sort_values(by='importance', ascending=False).reset_index(drop=True)

features_df_sorted

features_df_sorted.to_csv("./data/5_csv_dataframes/ctype_condition_multilabel_features_v3.csv", index=False)
