==> Part6.integration_evaluation_metrics.ipynb <==
import scanpy as sc
import hdf5plugin
from scalex.metrics import batch_entropy_mixing_score
from scalex.metrics import silhouette_score
import pandas as pd
import scib

from rpy2.robjects import pandas2ri
pandas2ri.activate()

### Concatenated dataset

# H5AD = "./data/4_concatenated_h5ad/lung_cancer_concat_data_v3.h5ad"

### SCANORAMA integrated datasets

# H5AD_INTEGRATED = "./data/6_integrated_h5ads/scanorama/lung_cancer_scanorama_all_genes.h5ad"

# H5AD_INTEGRATED = "./data/6_integrated_h5ads/scanorama/lung_cancer_scanorama_3000_hvgs.h5ad"

# H5AD_INTEGRATED = "./data/6_integrated_h5ads/scanorama/lung_cancer_scanorama_top_500_mlc_genes.h5ad"

# H5AD_INTEGRATED = "./data/6_integrated_h5ads/scanorama/lung_cancer_scanorama_top_300_mlc_genes.h5ad"

### HARMONY integrated datasets

# H5AD_INTEGRATED = "./data/6_integrated_h5ads/lung_cancer_harmony_all_genes.h5ad"

# H5AD_INTEGRATED = "./data/6_integrated_h5ads/lung_cancer_harmony_3000_hvgs.h5ad"

H5AD_INTEGRATED = "./data/6_integrated_h5ads/lung_cancer_harmony_top_500_mlc_genes.h5ad"

# H5AD_INTEGRATED = "./data/6_integrated_h5ads/lung_cancer_harmony_top_300_mlc_genes.h5ad"

adata_concat = sc.read_h5ad(H5AD)
sc.tl.pca(adata_concat)
adata_concat

# adata_scanorama_full = sc.read_h5ad(H5AD_INTEGRATED)
# adata_scanorama_full

adata_harmony_full = sc.read_h5ad(H5AD_INTEGRATED)
adata_harmony_full

### kBET for concat data

scib.me.kBET(adata_concat, batch_key="batch", label_key="cell_type", type_="full", embed="X_pca")

### kBET for Scanorama data

scib.me.kBET(adata_scanorama_full, batch_key="batch", label_key="cell_type", type_="embed", embed="Scanorama")

### kBET for Harmony data

scib.me.kBET(adata_harmony_full, batch_key="batch", label_key="cell_type", type_="embed", embed="X_pca_harmony")

### Silhouette batch for concat data

scib.me.silhouette_batch(adata_concat, batch_key="batch", label_key="cell_type", embed="X_pca")

### Silhouette batch for Scanorama data

scib.me.silhouette_batch(adata_scanorama_full, batch_key="batch", label_key="cell_type", embed="Scanorama")

### Silhouette batch for Harmony data

scib.me.silhouette_batch(adata_harmony_full, batch_key="batch", label_key="cell_type", embed="X_pca_harmony")
